Scaling smaller icons to 40px instead of previous 38...

+++ UIImage+RSRoundCorners.m
@@ -11,21 +11,21 @@
 
 void addRoundedRectToPath(CGContextRef context, CGRect rect, CGFloat ovalWidth, CGFloat ovalHeight, CGFloat radius)
 {
-    if (ovalWidth == 0 || ovalHeight == 0) {
+    if (ovalWidth == 0.0 || ovalHeight == 0.0) {
         CGContextAddRect(context, rect);
         return;
     }
     CGContextSaveGState(context);
     CGContextTranslateCTM (context, CGRectGetMinX(rect), CGRectGetMinY(rect));
     CGContextScaleCTM (context, ovalWidth, ovalHeight);
-    float fw = CGRectGetWidth (rect) / ovalWidth;
-    float fh = CGRectGetHeight (rect) / ovalHeight;
-    CGContextMoveToPoint(context, fw, fh/2);
+    CGFloat fw = CGRectGetWidth (rect) / ovalWidth;
+    CGFloat fh = CGRectGetHeight (rect) / ovalHeight;
+    CGContextMoveToPoint(context, fw, fh/2.0);
     
-    CGContextAddArcToPoint(context, fw, fh, fw/2, fh, radius);
-    CGContextAddArcToPoint(context, 0, fh, 0, fh/2, radius);
-    CGContextAddArcToPoint(context, 0, 0, fw/2, 0, radius);
-    CGContextAddArcToPoint(context, fw, 0, fw, fh/2, radius); 
+    CGContextAddArcToPoint(context, fw, fh, fw/2.0, fh, radius);
+    CGContextAddArcToPoint(context, 0.0, fh, 0.0, fh/2.0, radius);
+    CGContextAddArcToPoint(context, 0.0, 0.0, fw/2.0, 0, radius);
+    CGContextAddArcToPoint(context, fw, 0.0, fw, fh/2.0, radius); 
 
     CGContextClosePath(context);
     CGContextRestoreGState(context);
@@ -41,12 +41,12 @@ void addRoundedRectToPath(CGContextRef context, CGRect rect, CGFloat ovalWidth, 
     CGContextRef context = CGBitmapContextCreate(NULL, w, h, 8, 4 * w, colorSpace, kCGImageAlphaPremultipliedFirst);
     
     CGContextBeginPath(context);
-    CGRect rect = CGRectMake(0, 0, w, h);
+    CGRect rect = CGRectMake(0.0, 0.0, w, h);
     addRoundedRectToPath(context, rect, 4.0, 4.0, radius);
     CGContextClosePath(context);
     CGContextClip(context);
     
-    CGContextDrawImage(context, CGRectMake(0, 0, w, h), self.CGImage);
+    CGContextDrawImage(context, CGRectMake(0.0, 0.0, w, h), self.CGImage);
     
     CGImageRef imageMasked = CGBitmapContextCreateImage(context);
     CGContextRelease(context);
@@ -57,7 +57,7 @@ void addRoundedRectToPath(CGContextRef context, CGRect rect, CGFloat ovalWidth, 
 
 - (UIImage*)imageScaledToSize:(CGSize)size {
     UIGraphicsBeginImageContext(size);
-    [self drawInRect:CGRectMake(0, 0, size.width, size.height)];
+    [self drawInRect:CGRectMake(0.0, 0.0, size.width, size.height)];
     UIImage *image = UIGraphicsGetImageFromCurrentImageContext();
     UIGraphicsEndImageContext();
     return image;
